Setting up an offline coupler for the ISOMIP experiments
The coupler is made up of a driver script (coupledISOMIPRun.py)
that alternately runs MPAS-O and a script that computes new
temperature, salinity and mass fluxes (recomputeISOMIPMeltFluxes.py).
The driver can be used to select the number of time steps, the duration
of the coupling time step in hours (an integer), whether this is an "init"
run (as opposed to a restart) and the folder for the run where the
initial condition has been set up.

A template namelist file is required with entries @doRestart,
@startTime and @coupleTime entries defined, so that these can
be replaced depending on whether this is an "init" or a "restart"
run.