Adding an example namelist template to the ISOMIP coupler
The template shows how the @doRestart, @startTime and @coupleTime
fields are defined, and also typical namelist values for other
parameters used in the ISOMIP experiments.  The values @doRestart
and @startTime will be replaced by the driver as appropriate for
an "init" or a "restart" run.  @coupleTime sets both the duration
of a single MPAS-O run and the frequency of at which output and
restart files are written, and thus appears in several places in
the template.