Driver writes output to a single file; Cleanup of flux calculations
The output is concatinated into a single file using NCO's
ncrcat tool.  (This is slow but the result is more useful
for viz.)  Output steps that are kept are appended onto
the end of the file output.nc

The algorithm used to compute fluxes is unaltered (or rather
altered and then restored after careful thought) but the variable
names have been cleaned up and made more consistent with the
literature, especially Jenkins et al. (2001)