Add script for extracting transects
This merge adds a script, paraview_vtk_transect_extractor.py,
that can be used to extract data along transects specified by
a NetCDF file produced using the MaskCreator tool.

The syntax is similar to paraview_vtk_field_extractor.py, except
that only fields on cells are currently supported and the following
new flags have been added:
--transects_file: a NetCDF file produce by the MaskCreator that
contains one or more transects
--transect_dim: the depth dimension for transects (e.g. nVertLevels
in MPAS-Ocean)
--min_level_cell: a cell index array specifying the first valid
index in each column (0 by default)
--max_level_cell: a cell index array specifiying the last valid
index in each column (transect_dim-1 by default), e.g. maxLevelCell
in MPAS-Ocean
--layer_thickness: the name of the field (optionally with a
minus sign in front) where layer thicknesses are stored, used to
compute the zInterface field that can be used in Paraview as
the vertical coordinate for transects.
--z_min: the name of a field (optionally with a minus sign in front)
specifying the depth of the lower interface (in index space) of the
first valid layer on cells, used to compute zInterface
--z_max: the name of a field (optionally with a minus sign in front)
specifying the depth of the upper interface (in index space) of the
last valid layer on cells, used to compute zInterface

The script averages layerThickness from cell centers to edges (or used
the layerThickness of the valid cell for edges and layers where only
neighboring cell is valid, e.g. at stair steps in the topography).
Then, a cumulative sum of layer thickness is computed starting at either
z_min or z_max (z_min is used if both are specified) and stored in
zInterface for use in Paraview.