Combine topo and fixed depth extractors
The new script, paraview_vtk_field_at_depth_extractor.py, can either
extract fields at a fixed depth (--depth) or given a field on cells
that specifies the vertical index (--field_level_cell).  The latter
is used to extract fields in the last level above (or first
level below) topography.

The topography itself is also extracted into the fields zInterfaceTop
and zInterfaceBot (the top and bottom, respectively, of the level
from which fields were extracted).  Importantly, zInterfaceTop and
zInterfaceBot are not the same as bottomDepth at edges but instead
show how topography is represented in MPAS at edges.